+++ Scenes/ChatMenu/chat.gd
@@ -36,7 +36,9 @@ func _gui_input(event: InputEvent):
 		event = event as InputEventMouse
 		if event.button_mask != MouseButton.MOUSE_BUTTON_LEFT or not event.is_pressed():
 			return
-		DictEditMenu.close()
+		if DictEditMenu.is_open():
+			DictEditMenu.close()
+			accept_event()
 
 static func on_chat_display_child_entered_tree(node: Node, chat_body: VBoxContainer):
 	if node is Separator: