+++ Scenes/ChatMenu/transmission_entry.gd
@@ -18,8 +18,6 @@ var has_image: bool = false
 
 @onready var context_buttons: HFlowContainer = $Body/Corpus/ContextButtons
 
-var unknown_button = preload("res://Scenes/Common/unknown_signal_button.tscn")
-
 func ready():
 	transmission_node.get_popup().id_pressed.connect(transmit_pressed)
 	callsign_node.self_modulate = Main.get_callsign_color(sender)
@@ -56,28 +54,6 @@ func refresh():
 func set_message_text(new_text: String):
 	message_node.text = new_text
 	is_clickable = false
-	for but in context_buttons.get_children():
-		if but is UnknownSignalButton:
-			but.queue_free()
-	# failsafe cap of 8 unknown signals just in case
-	# shouldn't happen normally and defining them will show the rest
-	# ergo a setting isn't warranted
-	var unknown_cap: int = 8
-	var unknowns: Array[int] = []
-	for sig in message:
-		if unknown_cap <= 0:
-			break
-		if sig >= 0:
-			continue
-		if unknowns.has(sig):
-			continue
-		if not DictionaryHandler.word_keys.has(sig):
-			var new_button: UnknownSignalButton = unknown_button.instantiate()
-			new_button.text = DictionaryHandler.get_or_default_signal_name(sig)
-			new_button.sig = sig
-			context_buttons.add_child(new_button)
-			unknown_cap -= 1
-			unknowns.append(sig)
 
 func _on_etc_button_toggled(_toggled_on):
 	collapsed = not _toggled_on
@@ -104,6 +80,7 @@ func _on_hover_change(hovering: bool) -> void:
 	timeago_node.visible = hovering
 	hover_node.visible = hovering
 	is_hovering = hovering
+	_handle_clickables()
 
 func _on_set_etc_visibility(visiblity: bool) -> void:
 	etc_button_node.visible = visiblity
@@ -118,14 +95,17 @@ func _on_message_meta_clicked(meta: Variant) -> void:
 var is_hovering: bool = false
 var is_clickable: bool = false
 
-func _input(event):
-	if event.is_action_pressed("click_signal_modifier") and is_hovering and not is_clickable:
+func _handle_clickables():
+	if Input.is_action_pressed("click_signal_modifier") and is_hovering and not is_clickable:
 		request_rewrite(true)
 		if DisplayServer.cursor_get_shape() == DisplayServer.CURSOR_IBEAM:
 			DisplayServer.cursor_set_shape(DisplayServer.CURSOR_POINTING_HAND)
 		is_clickable = true
-	if (event.is_action_released("click_signal_modifier") or not is_hovering) and is_clickable:
+	if (not Input.is_action_pressed("click_signal_modifier") or not is_hovering) and is_clickable:
 		request_rewrite(false)
 		if DisplayServer.cursor_get_shape() == DisplayServer.CURSOR_POINTING_HAND:
 			DisplayServer.cursor_set_shape(DisplayServer.CURSOR_IBEAM)
 		is_clickable = false
+
+func _input(_event):
+	_handle_clickables()